+++ src/app/modules/user-student/student-home/components/student-posts/student-posts.component.ts
@@ -5,11 +5,9 @@ import { Post } from 'src/models';
 import { StudentGetPosts } from '../../store/actions';
 import { LoaderState } from '../../store/loader.state';
 import { StudentPostsState } from '../../store/student-post-state';
-import { COMMA, ENTER } from '@angular/cdk/keycodes';
 
 import { MatDialog } from '@angular/material/dialog';
-import { MAT_DIALOG_DATA } from '@angular/material/dialog';
-import { MatChipInputEvent } from '@angular/material/chips';
+import { PostPopupComponent } from '../popup-post/popup-post.component';
 @Component({
   selector: 'app-student-posts',
   templateUrl: './student-posts.component.html',
@@ -44,63 +42,5 @@ export class StudentPostsComponent implements OnInit {
     dialogRef.afterClosed().subscribe(result => {
       console.log(`Dialog result: ${result}`);
     });
-  }
-}
-@Component({
-  selector: 'app-popup-post.component.html',
-  templateUrl: './../popup-post/popup-post.component.html',
-  styleUrls: ['./../popup-post/popup-post.component.css']
-})
-export class PostPopupComponent implements OnInit{
-  showAddTasks!: boolean;
-  mediaUrl?: boolean = false;
-  visible = true;
-  selectable = true;
-  removable = true;
-  addOnBlur = true;
-  readonly separatorKeysCodes: number[] = [ENTER, COMMA];
-
-  //todo links from database
-  links = [{
-    url: 'https://material.angular.io/components/badge/overview'
-  },
-  {
-    url: 'https://material.angular.io/components/badge/overview'
-  }
-  ];
-
-  constructor(
-    @Inject(MAT_DIALOG_DATA) public data: Post,
-  ) {}
-
-  ngOnInit(): void {
-    if (this.data.mediaUrl.length !== 0) {
-     this.mediaUrl = !this.mediaUrl;
-    }
-  }
-
-  addTasks() {
-    this.showAddTasks = !this.showAddTasks;
-  }
-
-  add(event: MatChipInputEvent): void {
-    const input = event.input;
-    const value = event.value;
-
-    if ((value || '').trim()) {
-      this.links.push({ url: value.trim() });
-    }
-
-    if (input) {
-      input.value = '';
-    }
-  }
-
-  remove(link: any): void {
-    const index = this.links.indexOf(link);
-
-    if (index >= 0) {
-      this.links.splice(index, 1);
-    }
   }
 }
